WIP - refacto contexts
