+++ src/context/AuthContext/logic.js
@@ -1,0 +1,37 @@
+import { auth, db } from "../../firebase";
+import {
+  createUserWithEmailAndPassword,
+  onAuthStateChanged,
+  signInWithEmailAndPassword,
+  signOut,
+} from "firebase/auth";
+import { setDoc, doc } from "firebase/firestore";
+import { useEffect, useState } from "react";
+
+export const useAuth = () => {
+  const [user, setUser] = useState({});
+
+  useEffect(() => {
+    const unsubscribe = onAuthStateChanged(auth, (currentUser) => {
+      setUser(currentUser);
+    });
+    return () => {
+      unsubscribe();
+    };
+  }, []);
+  const signUp = (email, password) => {
+    createUserWithEmailAndPassword(auth, email, password);
+    setDoc(doc(db, "users", email), {
+      savedMovies: [],
+    });
+  };
+
+  const logIn = (email, password) => {
+    return signInWithEmailAndPassword(auth, email, password);
+  };
+
+  const logOut = () => {
+    return signOut(auth);
+  };
+  return { logIn, logOut, signUp, user };
+};
