+++ src/context/MoviesContext/logic.js
@@ -1,9 +1,8 @@
-import React, { createContext, useState, useEffect, useContext } from "react";
+import { useEffect, useState } from "react";
+import { requests } from "../../Requests";
 import axios from "axios";
-import { requests } from "../Requests";
-export const MoviesContext = createContext({});
 
-export const MoviesProvider = ({ children }) => {
+export const useMovies = () => {
   const [showMovieInfosModal, setShowMovieInfosModal] = useState(false);
   const [selectedMovie, setSelectedMovie] = useState({});
   const [moviesPopular, setMoviesPopular] = useState([]);
@@ -31,31 +30,18 @@ export const MoviesProvider = ({ children }) => {
     };
     getmoviesPopular();
   }, []);
-  if (isLoading) {
-    return <span>En cours de chargement...</span>;
-  } else
-    return (
-      <MoviesContext.Provider
-        value={{
-          showMovieInfosModal,
-          setShowMovieInfosModal,
-          moviesPopular,
-          setMoviesPopular,
-          moviesTopRated,
-          setMoviesTopRated,
-          moviesTrending,
-          setMoviesTrending,
-          selectedMovie,
-          setSelectedMovie,
-        }}
-      >
-        {children}
-      </MoviesContext.Provider>
-    );
-};
 
-export default MoviesProvider;
-
-export const useMoviesContext = () => {
-  return useContext(MoviesContext);
+  return {
+    showMovieInfosModal,
+    setShowMovieInfosModal,
+    moviesPopular,
+    setMoviesPopular,
+    moviesTopRated,
+    setMoviesTopRated,
+    moviesTrending,
+    setMoviesTrending,
+    selectedMovie,
+    setSelectedMovie,
+    isLoading,
+  };
 };
